fetch/metadata/generated/element-img.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-img.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, sourceAttr, attributes) {
    const image = document.createElement('img');

    for (const [ name, value ] of Object.entries(attributes)) {
      image.setAttribute(name, value);
    }

    return new Promise((resolve) => {
        image.setAttribute(sourceAttr, url);
        image.onload = image.onerror = resolve;
      });
  }

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - srcset - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameOrigin']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - Cross-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameOrigin']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - srcset - Cross-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameSite']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - Cross-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameSite']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - srcset - Cross-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'crossSite']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - Cross-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'crossSite']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - srcset - Cross-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameOrigin']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - src - Same-Origin -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameOrigin']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - srcset - Same-Origin -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameSite']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - src - Same-Origin -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameSite']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - srcset - Same-Origin -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - Same-Origin -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - srcset - Same-Origin -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameOrigin']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - src - Same-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameOrigin']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - srcset - Same-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameSite']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - src - Same-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameSite']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - srcset - Same-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'crossSite']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - Same-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'crossSite']),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - srcset - Same-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin']),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - src - HTTPS downgrade-upgrade, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'no-cors');
        });
  }, 'sec-fetch-mode - src - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'src',
        {"crossorigin": ""}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - src - attributes: crossorigin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'src',
        {"crossorigin": "anonymous"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - src - attributes: crossorigin=anonymous');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'src',
        {"crossorigin": "use-credentials"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - src - attributes: crossorigin=use-credentials');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'no-cors');
        });
  }, 'sec-fetch-mode - srcset - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'srcset',
        {"crossorigin": ""}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - srcset - attributes: crossorigin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'srcset',
        {"crossorigin": "anonymous"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - srcset - attributes: crossorigin=anonymous');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'srcset',
        {"crossorigin": "use-credentials"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'cors');
        });
  }, 'sec-fetch-mode - srcset - attributes: crossorigin=use-credentials');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'image');
        });
  }, 'sec-fetch-dest - src - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'image');
        });
  }, 'sec-fetch-dest - srcset - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'src',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - src - no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, []),
        'srcset',
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - srcset - no attributes');
  </script>
  </body>
</html>
